requirements.txt
ansicon==1.89.0
asgiref==3.7.2
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
blessed==1.20.0
boto3==1.26.151
botocore==1.29.151
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.6
colored==1.4.4
cursor==1.3.5
Django==4.2.1
exceptiongroup==1.1.1
fade==0.0.9
h11==0.14.0
idna==3.4
image==1.5.33
inquirer==3.1.3
jinxed==1.2.0
jmespath==1.0.1
outcome==1.2.0
packaging==23.1
Pillow==9.5.0
prettytable==3.7.0
pycparser==2.21
pyfiglet==0.8.post1
PySocks==1.7.1
pystyle==2.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
readchar==4.0.5
requests==2.27.1
s3transfer==0.6.1
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
tls-client==0.2.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.6.2
tzdata==2023.3
undetected-chromedriver==3.1.7
urllib3==1.26.16
wcwidth==0.2.6
webdriver-manager==3.8.6
websockets==11.0.3
wsproto==1.2.0
